feat: moved to components and refactoring
